ocr-lib/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ocrlib"
version = "0.1.0"
description = "OCR label scanning library — extract fields from device label photos"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
